Might rework interpreter in python
After looking at how this project might be implemented, I determined
that it might be best to write the interpreter in python instead of C.
This is due to the fact that , 1. the interface will be in python so it
will be easier to transfer data between those parts of the application
2. Makes it easier for anyone who might be working on the project
because they wont need sufficiency in 2 languages 3. It will save
memory and compile time since a bridge won’t need to be created